Marcha Atrás No se puede incluir varios archivos javascript y es tontería hacerlo, porque si los separos, desde diferentes archivos invocan al mismo, cargándolo dos veces.
