+++ funciones/principal.js
@@ -4,6 +4,7 @@ import configurar_idioma from "./opciones.js";
 import cambiar_idioma from "./idioma.js";
 import cambiar_tema from "./tema.js";
 
+const botón_idioma = document.getElementById("botón_idioma");
 const botón_tema = document.getElementById("botón_tema");
 const nodo_idioma = document.getElementById("lista_idioma");
 const nodo_estilo = document.getElementById("tema");
@@ -19,4 +20,14 @@ predeterminado.then(datos => {
 	cambiar_idioma(datos.idioma);
 	// Cambiamos al tema por defecto.
 	if (!datos.modo_oscuro) cambiar_tema(nodo_estilo, botón_tema);
+});
+
+// Botón para cambiar el idioma.
+botón_idioma.addEventListener("change", () => {
+	cambiar_idioma(botón_idioma.value);
+});
+
+// Botón para cambiar el tema.
+botón_tema.addEventListener("click", () => {
+	cambiar_tema(nodo_estilo, botón_tema);
 });
