+++ index.html
@@ -38,11 +38,8 @@
 	</main>
 	<!-- Información legal y menciones -->
 	<footer id="pie"></footer>
-	<!-- Importante ponerlo abajo para mejor carga -->
-	<script type="module" src="funciones/principal.js"></script> <!-- Carga visual -->
-	<script type="module" src="funciones/"></script> <!-- Carga inicial 1 -->
-	<script type="module" src="funciones/"></script> <!-- Carga inicial 2 -->
-	<script type="module" src="funciones/eventos.js"></script>
+	<!-- Importante ponerlo abajo para mejorar la carga -->
+	<script type="module" src="funciones/principal.js"></script>
 </body>
 
 </html>
